Small changes to the prediction interface of the additive model, because it wasn't using the nan flag yet.
